add table comment and  consummate handle error

+++ src/main/java/life/wpx1997/community/dto/QuestionDTO.java
@@ -5,16 +5,16 @@ import lombok.Data;
 
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String accountId;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
-    private Integer viewCount;
-    private Integer commentCount;
-    private Integer likeCount;
+    private Long creator;
+    private Long viewCount;
+    private Long commentCount;
+    private Long likeCount;
     private User user;
 }
